created moduler routes and fixed bug:handle reviews related to deleted listing
